+++ example/toolloop/main.go
@@ -2,11 +2,12 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 	"strings"
 	"time"
+
+	"github.com/gtkit/json"
 
 	"github.com/gtkit/go-llm-provider/provider"
 )